requirements.txt
# Core scientific stack
numpy==1.26.4
pandas
scipy==1.16.3
matplotlib
seaborn
Pillow

# Machine learning
scikit-learn==1.7.2
scikit-image~=0.25
cellpose==3.1.1.2
tensorflow==2.16.2
keras==3.10  # 3.11 not compatible
torch==2.5.1
torchvision>=0.20.1
typing_extensions==4.15.0

# Computer vision
opencv-python
opencv-contrib-python==4.10.0.84
opencv_python_headless==4.11.0.86
tifffile==2024.9.20
sam2
huggingface_hub

# Visualization and web app
plotly==6.3.1
streamlit==1.40.0
streamlit-image-coordinates==0.1.9   # last pre-2025 release, works with 1.40
